+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/client/tunnelling/HiveMetaStoreClientSupplierFactory.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2018 Expedia, Inc.
+ * Copyright (C) 2016-2019 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -35,7 +35,7 @@ public class HiveMetaStoreClientSupplierFactory {
     this(new DefaultMetaStoreClientFactory());
   }
 
-  HiveMetaStoreClientSupplier newInstance(HiveConf hiveConf, String name, int reconnectionRetries) {
-    return new HiveMetaStoreClientSupplier(factory, hiveConf, name, reconnectionRetries);
+  HiveMetaStoreClientSupplier newInstance(HiveConf hiveConf, String name, int reconnectionRetries, int connectionTimeout) {
+    return new HiveMetaStoreClientSupplier(factory, hiveConf, name, reconnectionRetries, connectionTimeout);
   }
 }