+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/mapping/model/DatabaseMappingImpl.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2019 Expedia, Inc.
+ * Copyright (C) 2016-2019 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -110,7 +110,6 @@ public class DatabaseMappingImpl implements DatabaseMapping {
       }
     }
     return table;
-
   }
 
   @Override
@@ -485,8 +484,7 @@ public class DatabaseMappingImpl implements DatabaseMapping {
   }
 
   @Override
-  public long getTimeout() {
-    return metaStoreMapping.getTimeout();
+  public long getLatency() {
+    return metaStoreMapping.getLatency();
   }
-
 }