+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/mapping/model/MetaStoreMapping.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2019 Expedia, Inc.
+ * Copyright (C) 2016-2019 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -63,5 +63,5 @@ public interface MetaStoreMapping extends Closeable {
   void createDatabase(Database database)
     throws AlreadyExistsException, InvalidObjectException, MetaException, TException;
 
-  long getTimeout();
+  long getLatency();
 }