+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/mapping/model/MetaStoreMappingImpl.java
@@ -40,7 +40,7 @@ class MetaStoreMappingImpl implements MetaStoreMapping {
   private final CloseableThriftHiveMetastoreIface client;
   private final AccessControlHandler accessControlHandler;
   private final String name;
-  private final long timeout;
+  private final long latency;
 
   private final ConnectionType connectionType;
 
@@ -50,13 +50,13 @@ class MetaStoreMappingImpl implements MetaStoreMapping {
       CloseableThriftHiveMetastoreIface client,
       AccessControlHandler accessControlHandler,
       ConnectionType connectionType,
-      long timeout) {
+      long latency) {
     this.databasePrefix = databasePrefix;
     this.name = name;
     this.client = client;
     this.accessControlHandler = accessControlHandler;
     this.connectionType = connectionType;
-    this.timeout = timeout;
+    this.latency = latency;
   }
 
   @Override
@@ -135,7 +135,7 @@ class MetaStoreMappingImpl implements MetaStoreMapping {
   }
 
   @Override
-  public long getTimeout() {
-    return timeout;
+  public long getLatency() {
+    return latency;
   }
 }