+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/mapping/service/requests/GetAllDatabasesRequest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2019 Expedia, Inc.
+ * Copyright (C) 2016-2019 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,28 +16,32 @@
 package com.hotels.bdp.waggledance.mapping.service.requests;
 
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.Callable;
+import java.util.function.BiFunction;
 
 import org.apache.thrift.TException;
 
 import com.hotels.bdp.waggledance.mapping.model.DatabaseMapping;
-import com.hotels.bdp.waggledance.mapping.service.DatabaseMappingUtils;
-import com.hotels.bdp.waggledance.util.Whitelist;
 
-public class GetAllDatabasesRequest implements Callable<List<?>> {
+public class GetAllDatabasesRequest implements RequestCallable<List<String>> {
 
   private final DatabaseMapping mapping;
-  private final Map<String, Whitelist> mappedDbByPrefix;
+  private final BiFunction<List<String>, DatabaseMapping, List<String>> filter;
 
-  public GetAllDatabasesRequest(DatabaseMapping mapping, Map<String, Whitelist> mappedDbByPrefix) {
+  public GetAllDatabasesRequest(
+      DatabaseMapping mapping,
+      BiFunction<List<String>, DatabaseMapping, List<String>> filter) {
     this.mapping = mapping;
-    this.mappedDbByPrefix = mappedDbByPrefix;
+    this.filter = filter;
   }
 
   @Override
   public List<String> call() throws TException {
     List<String> databases = mapping.getClient().get_all_databases();
-    return DatabaseMappingUtils.getMappedWhitelistedDatabases(databases, mapping, mappedDbByPrefix);
+    return filter.apply(databases, mapping);
+  }
+
+  @Override
+  public DatabaseMapping getMapping() {
+    return mapping;
   }
 }