+++ waggle-dance-core/src/test/java/com/hotels/bdp/waggledance/mapping/service/impl/MonitoredDatabaseMappingServiceTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2018 Expedia, Inc.
+ * Copyright (C) 2016-2019 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,6 +20,9 @@ import static org.junit.Assert.assertThat;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+
+import java.util.Arrays;
+import java.util.List;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -99,5 +102,12 @@ public class MonitoredDatabaseMappingServiceTest {
   public void close() throws Exception {
     service.close();
     verify(wrapped).close();
+  }
+
+  @Test
+  public void getDatabaseMappings() {
+    List<DatabaseMapping> databaseMappings = Arrays.asList(primaryMapping, otherMapping);
+    when(wrapped.getDatabaseMappings()).thenReturn(databaseMappings);
+    assertThat(service.getDatabaseMappings(), is(databaseMappings));
   }
 }