+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 ## TBD
+### Added
+* Configurable `latency` for each metastore in a Waggle Dance configuration.
+
 ### Fixed
 * Support for regex `mapped-databases` for MANUAL database resolution. See [#147](https://github.com/HotelsDotCom/waggle-dance/issues/147).
 * Avoid NPE when no elements are provided for `mapped-databases` in the configuration. See [#131](https://github.com/HotelsDotCom/waggle-dance/issues/131).
 * Support for running metastore queries in parallel.
 
 ### Changed
-* Updated `hotels-oss-parent` to 4.0.1 (was 2.3.5).
+* Updated `hotels-oss-parent` to 4.0.0 (was 2.3.5).
+* Added invocation log messages for `getPartitions` calls.
 
 ## [3.1.2] - 2019-01-11
 ### Changed